your mom

+++ src/add-metric.php
@@ -2,14 +2,16 @@
 
 include 'config.php';
 
-if (isset($_POST['weight']) && isset($_POST['height']) && isset($_POST['name'])){
+if (isset($_POST['weight']) && isset($_POST['height']) && isset($_POST['name']) && isset($_POST['gender']) && isset($_POST['age'])){
     $weight = $_POST['weight'];
     $height = $_POST['height'];
     $name = $_POST['name'];
+    $gender = $_POST['gender'];
+    $age = $_POST['age'];
     if ($name != "" && $height != 0 && $weight != 0){
         $executeFirst = "DELETE FROM metric";
         mysqli_query($conn, $executeFirst);
-        $sql = "INSERT INTO metric (name, weight,height) VALUES ('$name', '$weight','$height')";
+        $sql = "INSERT INTO metric (name, weight,height, gender, age) VALUES ('$name', '$weight','$height', '$gender', '$age')";
         $result = mysqli_query($conn, $sql);
         if ($result) {
             echo 1;
